requirements.txt
Flask>=3.1.3
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.1.0
PyMySQL>=1.1.2
cryptography>=46.0.5
pandas>=3.0.1
openpyxl>=3.1.5
requests>=2.32.5
python-docx>=1.2.0
celery>=5.6.2
redis>=6.4.0
celery[redis]
gunicorn>=25.1.0
deepdiff>=8.6.1
docling>=2.79.0
openai>=2.26.0